test_raw_api moved out of raw_api to avoid circular imports
