+++ test/integration/cleanup_buckets.py
@@ -8,9 +8,9 @@
 #
 ######################################################################
 
-from b2sdk.raw_api import cleanup_old_buckets
 from . import get_b2_auth_data
 from .test_large_files import BucketCleaner
+from .test_raw_api import cleanup_old_buckets
 
 if __name__ == '__main__':
     cleanup_old_buckets()
